requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
drf-nested-routers==0.93.3
pytz==2021.1
sqlparse==0.4.2
djangorestframework-simplejwt==5.0.0
